requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
mysql-connector-python==8.0.13
numpy==1.15.4
protobuf==3.6.1
python-dateutil==2.7.5
python-slugify==2.0.0
pytz==2018.7
requests==2.21.0
six==1.12.0
termcolor==1.1.0
Unidecode==1.0.23
urllib3==1.24.1